Add TestMerge

+++ src/main/java/test/Test.java
@@ -2,6 +2,8 @@ package test;
 
 import org.hibernate.Session;
 
+import java.util.Set;
+
 public class Test {
     public static void main(String[] args) {
         Session session = HibernateUtil.getSessionFactory().openSession();
@@ -9,6 +11,7 @@ public class Test {
         session.beginTransaction();
 
         Assignment a1 = new Assignment();
+        a1.setOwnerOid("123");
         a1.setId(10);
 
         AssignmentExtension ae1 = new AssignmentExtension();
@@ -30,18 +33,11 @@ public class Test {
 
         session.beginTransaction();
 
-        Assignment loaded = session.get(Assignment.class, new AssignmentId(10));
-//		if (loaded.getExtension() != null) {
-//			Set<ExtBoolean> booleans = loaded.getExtension().getBooleans();
-//			if (booleans.size() != 1) {
-//				throw new AssertionError("Unexpected size of booleans collection: " + booleans.size());
-//			}
-//			// temporary
-//			for (ExtBoolean extBoolean : booleans) {
-//				System.out.println("+++++++++++ deleting " + extBoolean);
-//				session.delete(extBoolean);
-//			}
-//		}
+        Assignment loaded = session.get(Assignment.class, new Assignment("123", 10));
+	    Set<ExtBoolean> booleans = loaded.getExtension().getBooleans();
+	    if (booleans.size() != 1) {
+		    throw new AssertionError("Unexpected size of booleans collection: " + booleans.size());
+	    }
         session.delete(loaded);
         session.getTransaction().commit();
